requirements.txt
fastapi
uvicorn
torch
scikit-learn
pydantic
numpy
pandas
python-multipart
rdkit-pypi